Live deployment

+++ vfun/initial_db_load.py
@@ -2,28 +2,38 @@ from pathlib import Path
 from .models import SportsHall
 import csv
 from django.contrib.gis.geos import Point
+from django.contrib.auth import authenticate
 
 sports_halls_csv = Path(__file__).resolve().parent / 'data' / 'sports_halls.csv'
 
 
 def run():
-    with open(sports_halls_csv) as file:
-        reader = csv.reader(file)
-        # pass the header
-        next(reader)
+    username = input('Enter your username:')
+    password = input('Enter your password:')
 
-        SportsHall.objects.all().delete()
+    user = authenticate(username=username, password=password)
 
-        for row in reader:
-            print(row)
-            lon = float(row[2])
-            lat = float(row[3])
+    if user is not None:
+        with open(sports_halls_csv) as file:
+            reader = csv.reader(file)
+            # pass the header
+            next(reader)
 
-            sports_hall = SportsHall(
-                hall_name=row[0],
-                address=row[1],
-                location=Point(lon, lat),
-                courts=int(row[4]),
-                phone_no=row[5],
-            )
-            sports_hall.save()
+            SportsHall.objects.all().delete()
+
+            for row in reader:
+                print(row)
+                lon = float(row[2])
+                lat = float(row[3])
+
+                sports_hall = SportsHall(
+                    hall_name=row[0],
+                    address=row[1],
+                    location=Point(lon, lat),
+                    courts=int(row[4]),
+                    phone_no=row[5],
+                    created_by=user.profile,
+                )
+                sports_hall.save()
+    else:
+        print("Username/Password incorrect, please try again.")
